meshbot/requirements.txt
google_api_python_client==2.56.0
google_auth_oauthlib==0.5.2
pandas==1.4.3
protobuf==4.21.5
python-dotenv==0.20.0
